pyproject.toml
[project]
name = "agentic-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.6",
    "langgraph-cli[inmem]>=0.4.23",
    "langsmith>=0.7.30",
    "python-dotenv>=1.2.2",
]
